// src/api/generated/schemas/getAiPricing200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { GetAiPricing200DataItemDescriptionEn } from './getAiPricing200DataItemDescriptionEn';
import type { GetAiPricing200DataItemDescriptionAr } from './getAiPricing200DataItemDescriptionAr';
import type { GetAiPricing200DataItemUpdatedById } from './getAiPricing200DataItemUpdatedById';
import type { GetAiPricing200DataItemCreatedAt } from './getAiPricing200DataItemCreatedAt';
import type { GetAiPricing200DataItemUpdatedAt } from './getAiPricing200DataItemUpdatedAt';

export type GetAiPricing200DataItem = {
  key: string;
  nameEn: string;
  nameAr: string;
  /** @nullable */
  descriptionEn: GetAiPricing200DataItemDescriptionEn;
  /** @nullable */
  descriptionAr: GetAiPricing200DataItemDescriptionAr;
  lenses: number;
  enabled: boolean;
  maxPromptChars: number;
  maxOutputTokens: number;
  sortOrder: number;
  /** @nullable */
  updatedById: GetAiPricing200DataItemUpdatedById;
  createdAt: GetAiPricing200DataItemCreatedAt;
  updatedAt: GetAiPricing200DataItemUpdatedAt;
};