// src/api/generated/schemas/getBillingSummary200AiFeaturesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { GetBillingSummary200AiFeaturesItemValueUsd } from './getBillingSummary200AiFeaturesItemValueUsd';

export type GetBillingSummary200AiFeaturesItem = {
  feature: string;
  actions: number;
  failures: number;
  refunds: number;
  costUsd: string;
  averageCostUsd: string;
  lensesCharged: number;
  /** @nullable */
  valueUsd: GetBillingSummary200AiFeaturesItemValueUsd;
};